+++ tests/en_test.go
@@ -4,7 +4,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/SerhiiCho/timeago/v3"
+	ago "github.com/SerhiiCho/timeago/v3"
+	"github.com/SerhiiCho/timeago/v3/internal/utils"
 )
 
 func TestParseEn(t *testing.T) {
@@ -12,50 +13,50 @@ func TestParseEn(t *testing.T) {
 		date time.Time
 		res  string
 	}{
-		{subMinutes(1), "1 minute ago"},
-		{subMinutes(2), "2 minutes ago"},
-		{subMinutes(5), "5 minutes ago"},
-		{subMinutes(9), "9 minutes ago"},
-		{subMinutes(10), "10 minutes ago"},
-		{subMinutes(11), "11 minutes ago"},
-		{subMinutes(20), "20 minutes ago"},
-		{subMinutes(21), "21 minutes ago"},
-		{subMinutes(22), "22 minutes ago"},
-		{subMinutes(30), "30 minutes ago"},
-		{subMinutes(31), "31 minutes ago"},
-		{subMinutes(59), "59 minutes ago"},
-		{subHours(1), "1 hour ago"},
-		{subHours(2), "2 hours ago"},
-		{subHours(9), "9 hours ago"},
-		{subHours(10), "10 hours ago"},
-		{subHours(11), "11 hours ago"},
-		{subHours(20), "20 hours ago"},
-		{subHours(21), "21 hours ago"},
-		{subHours(23), "23 hours ago"},
-		{subDays(1), "1 day ago"},
-		{subDays(2), "2 days ago"},
-		{subDays(4), "4 days ago"},
-		{subDays(5), "5 days ago"},
-		{subDays(6), "6 days ago"},
-		{subWeeks(1), "1 week ago"},
-		{subWeeks(2), "2 weeks ago"},
-		{subWeeks(3), "3 weeks ago"},
-		{subMonths(1), "1 month ago"},
-		{subMonths(2), "2 months ago"},
-		{subMonths(9), "9 months ago"},
-		{subMonths(11), "11 months ago"},
-		{subYears(1), "1 year ago"},
-		{subYears(2), "2 years ago"},
-		{subYears(21), "21 years ago"},
-		{subYears(31), "31 years ago"},
-		{subYears(100), "100 years ago"},
+		{utils.SubMinutes(1), "1 minute ago"},
+		{utils.SubMinutes(2), "2 minutes ago"},
+		{utils.SubMinutes(5), "5 minutes ago"},
+		{utils.SubMinutes(9), "9 minutes ago"},
+		{utils.SubMinutes(10), "10 minutes ago"},
+		{utils.SubMinutes(11), "11 minutes ago"},
+		{utils.SubMinutes(20), "20 minutes ago"},
+		{utils.SubMinutes(21), "21 minutes ago"},
+		{utils.SubMinutes(22), "22 minutes ago"},
+		{utils.SubMinutes(30), "30 minutes ago"},
+		{utils.SubMinutes(31), "31 minutes ago"},
+		{utils.SubMinutes(59), "59 minutes ago"},
+		{utils.SubHours(1), "1 hour ago"},
+		{utils.SubHours(2), "2 hours ago"},
+		{utils.SubHours(9), "9 hours ago"},
+		{utils.SubHours(10), "10 hours ago"},
+		{utils.SubHours(11), "11 hours ago"},
+		{utils.SubHours(20), "20 hours ago"},
+		{utils.SubHours(21), "21 hours ago"},
+		{utils.SubHours(23), "23 hours ago"},
+		{utils.SubDays(1), "1 day ago"},
+		{utils.SubDays(2), "2 days ago"},
+		{utils.SubDays(4), "4 days ago"},
+		{utils.SubDays(5), "5 days ago"},
+		{utils.SubDays(6), "6 days ago"},
+		{utils.SubWeeks(1), "1 week ago"},
+		{utils.SubWeeks(2), "2 weeks ago"},
+		{utils.SubWeeks(3), "3 weeks ago"},
+		{utils.SubMonths(1), "1 month ago"},
+		{utils.SubMonths(2), "2 months ago"},
+		{utils.SubMonths(9), "9 months ago"},
+		{utils.SubMonths(11), "11 months ago"},
+		{utils.SubYears(1), "1 year ago"},
+		{utils.SubYears(2), "2 years ago"},
+		{utils.SubYears(21), "21 years ago"},
+		{utils.SubYears(31), "31 years ago"},
+		{utils.SubYears(100), "100 years ago"},
 	}
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.Reconfigure(timeago.Config{Language: "en"})
+			ago.Reconfigure(ago.Config{Language: ago.LangEn})
 
-			res, err := timeago.Parse(tc.date)
+			res, err := ago.Parse(tc.date)
 
 			if err != nil {
 				test.Errorf("Error must be nil, but got %v instead", err)
@@ -73,24 +74,24 @@ func TestParseEnWithSeconds(t *testing.T) {
 		date time.Time
 		res  []string
 	}{
-		{subSeconds(0), []string{"0 seconds ago", "1 second ago"}},
-		{subSeconds(1), []string{"1 second ago", "2 seconds ago"}},
-		{subSeconds(2), []string{"2 seconds ago", "3 seconds ago"}},
-		{subSeconds(9), []string{"9 seconds ago", "10 seconds ago"}},
-		{subSeconds(10), []string{"10 seconds ago", "11 seconds ago"}},
-		{subSeconds(11), []string{"11 seconds ago", "12 seconds ago"}},
-		{subSeconds(20), []string{"20 seconds ago", "21 seconds ago"}},
-		{subSeconds(21), []string{"21 seconds ago", "22 seconds ago"}},
-		{subSeconds(22), []string{"22 seconds ago", "23 seconds ago"}},
-		{subSeconds(30), []string{"30 seconds ago", "31 seconds ago"}},
-		{subSeconds(59), []string{"59 seconds ago", "1 minute ago"}},
+		{utils.SubSeconds(0), []string{"0 seconds ago", "1 second ago"}},
+		{utils.SubSeconds(1), []string{"1 second ago", "2 seconds ago"}},
+		{utils.SubSeconds(2), []string{"2 seconds ago", "3 seconds ago"}},
+		{utils.SubSeconds(9), []string{"9 seconds ago", "10 seconds ago"}},
+		{utils.SubSeconds(10), []string{"10 seconds ago", "11 seconds ago"}},
+		{utils.SubSeconds(11), []string{"11 seconds ago", "12 seconds ago"}},
+		{utils.SubSeconds(20), []string{"20 seconds ago", "21 seconds ago"}},
+		{utils.SubSeconds(21), []string{"21 seconds ago", "22 seconds ago"}},
+		{utils.SubSeconds(22), []string{"22 seconds ago", "23 seconds ago"}},
+		{utils.SubSeconds(30), []string{"30 seconds ago", "31 seconds ago"}},
+		{utils.SubSeconds(59), []string{"59 seconds ago", "1 minute ago"}},
 	}
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.Reconfigure(timeago.Config{Language: "en"})
+			ago.Reconfigure(ago.Config{Language: ago.LangEn})
 
-			res, err := timeago.Parse(tc.date)
+			res, err := ago.Parse(tc.date)
 
 			if err != nil {
 				test.Errorf("Error must be nil, but got %v instead", err)