+++ timeago_test.go
@@ -3,6 +3,8 @@ package timeago
 import (
 	"testing"
 	"time"
+
+	"github.com/SerhiiCho/timeago/v3/internal/utils"
 )
 
 func TestParseFunctionCanExceptTimestamp(t *testing.T) {
@@ -10,21 +12,21 @@ func TestParseFunctionCanExceptTimestamp(t *testing.T) {
 		date int
 		res  string
 	}{
-		{timestampFromPastDate(time.Minute), "1 minute ago"},
-		{timestampFromPastDate(time.Minute * 5), "5 minutes ago"},
-		{timestampFromPastDate(time.Hour), "1 hour ago"},
-		{timestampFromPastDate(time.Hour * 3), "3 hours ago"},
-		{timestampFromPastDate(time.Hour * 5), "5 hours ago"},
-		{timestampFromPastDate(time.Hour * 24), "1 day ago"},
-		{timestampFromPastDate(time.Hour * 24 * 2), "2 days ago"},
-		{timestampFromPastDate(time.Hour * 24 * 3), "3 days ago"},
-		{timestampFromPastDate(time.Hour * 24 * 4), "4 days ago"},
-		{timestampFromPastDate(time.Hour * 24 * 5), "5 days ago"},
-		{timestampFromPastDate(time.Hour * 24 * 6), "6 days ago"},
-		{timestampFromPastDate(time.Hour * 24 * 7), "1 week ago"},
+		{utils.UnixFromPastDate(time.Minute), "1 minute ago"},
+		{utils.UnixFromPastDate(time.Minute * 5), "5 minutes ago"},
+		{utils.UnixFromPastDate(time.Hour), "1 hour ago"},
+		{utils.UnixFromPastDate(time.Hour * 3), "3 hours ago"},
+		{utils.UnixFromPastDate(time.Hour * 5), "5 hours ago"},
+		{utils.UnixFromPastDate(time.Hour * 24), "1 day ago"},
+		{utils.UnixFromPastDate(time.Hour * 24 * 2), "2 days ago"},
+		{utils.UnixFromPastDate(time.Hour * 24 * 3), "3 days ago"},
+		{utils.UnixFromPastDate(time.Hour * 24 * 4), "4 days ago"},
+		{utils.UnixFromPastDate(time.Hour * 24 * 5), "5 days ago"},
+		{utils.UnixFromPastDate(time.Hour * 24 * 6), "6 days ago"},
+		{utils.UnixFromPastDate(time.Hour * 24 * 7), "1 week ago"},
 	}
 
-	Reconfigure(Config{Language: "en"})
+	Reconfigure(Config{Language: LangEn})
 
 	for _, tc := range cases {
 		t.Run(tc.res, func(t *testing.T) {
@@ -46,20 +48,20 @@ func TestParseFunctionCanExceptTimePackage(t *testing.T) {
 		date time.Time
 		res  string
 	}{
-		{time.Now().Add(-time.Minute), "1 minute ago"},
-		{time.Now().Add(-time.Minute * 2), "2 minutes ago"},
-		{time.Now().Add(-time.Minute * 3), "3 minutes ago"},
-		{time.Now().Add(-time.Minute * 4), "4 minutes ago"},
-		{time.Now().Add(-time.Minute * 5), "5 minutes ago"},
-		{time.Now().Add(-time.Minute * 6), "6 minutes ago"},
-		{time.Now().Add(-time.Hour * 7), "7 hours ago"},
-		{time.Now().Add(-time.Hour * 8), "8 hours ago"},
-		{time.Now().Add(-time.Hour * 9), "9 hours ago"},
-		{time.Now().Add(-time.Hour * 10), "10 hours ago"},
-		{time.Now().Add(-time.Hour * 11), "11 hours ago"},
+		{utils.SubMinutes(1), "1 minute ago"},
+		{utils.SubMinutes(2), "2 minutes ago"},
+		{utils.SubMinutes(3), "3 minutes ago"},
+		{utils.SubMinutes(4), "4 minutes ago"},
+		{utils.SubMinutes(5), "5 minutes ago"},
+		{utils.SubMinutes(6), "6 minutes ago"},
+		{utils.SubHours(7), "7 hours ago"},
+		{utils.SubHours(8), "8 hours ago"},
+		{utils.SubHours(9), "9 hours ago"},
+		{utils.SubHours(10), "10 hours ago"},
+		{utils.SubHours(11), "11 hours ago"},
 	}
 
-	Reconfigure(Config{Language: "en"})
+	Reconfigure(Config{Language: LangEn})
 
 	for _, tc := range cases {
 		t.Run("Test for date "+tc.date.String(), func(t *testing.T) {
@@ -81,15 +83,15 @@ func TestParseFuncWillCalculateIntervalToFutureDate(t *testing.T) {
 		date time.Time
 		res  string
 	}{
-		{time.Now().Add(time.Minute * 2), "2 minutes"},
-		{time.Now().Add(time.Minute * 5), "5 minutes"},
-		{time.Now().Add(time.Minute * 10), "10 minutes"},
-		{time.Now().Add(time.Hour), "1 hour"},
-		{time.Now().Add(time.Hour * 24), "1 day"},
-		{time.Now().Add(time.Hour * 48), "2 days"},
+		{utils.AddMinutes(2), "2 minutes"},
+		{utils.AddMinutes(5), "5 minutes"},
+		{utils.AddMinutes(10), "10 minutes"},
+		{utils.AddHours(1), "1 hour"},
+		{utils.AddHours(24), "1 day"},
+		{utils.AddHours(48), "2 days"},
 	}
 
-	Reconfigure(Config{Language: "en"})
+	Reconfigure(Config{Language: LangEn})
 
 	for _, tc := range testCases {
 		t.Run("Test for date: "+tc.date.String(), func(t *testing.T) {